+++ api/models/stationpoint.go
@@ -4,11 +4,12 @@ import (
 	"time"
 )
 
-type StationPoint struct { // Points given to a station
-	ID			uint		`json:"id" gorm:"primary_key"`
+// Points given to a station
+type StationPoint struct {
+	ID			int64		`json:"id" gorm:"primary_key"`
 	CreatedAt	time.Time
 	UpdatedAt	time.Time
-	GroupID		uint		`json:"GroupID" gorm:"foreignKey:GroupID;index:idx_sp,unique"`
-	StationID	uint		`json:"StationID" gorm:"foreignKey:StationID;index:idx_sp,unique"`
-	Value		uint		`json:"value"`
+	GroupID		int64		`json:"GroupID" gorm:"foreignKey:GroupID;index:idx_sp,unique"`
+	StationID	int64		`json:"StationID" gorm:"foreignKey:StationID;index:idx_sp,unique"`
+	Value		int64		`json:"value"`
 }
